refactor cleanup
