+++ lib/character_shower/character_shower.dart
@@ -1,3 +1,4 @@
 export '../character_shower/data/data.dart';
 export '../character_shower/pages/pages.dart';
+export '../character_shower/providers/character_shower_providers.dart';
 export '../character_shower/widgets/widgets.dart';
